+++ src/gcovr/data_model/coverage.py
@@ -83,11 +83,7 @@ from .coverage_dict import (
     CoverageDict,
     LinesKeyType,
 )
-from .merging import (
-    DEFAULT_MERGE_OPTIONS,
-    GcovrMergeAssertionError,
-    MergeOptions,
-)
+from .merging import DEFAULT_MERGE_OPTIONS, MergeOptions
 from .stats import CoverageStat, DecisionCoverageStat, SummarizedStats
 
 LOGGER = logging.getLogger("gcovr")
@@ -96,6 +92,14 @@ GCOVR_DATA_SOURCES = "gcovr/data_sources"
 GCOVR_EXCLUDED = "gcovr/excluded"
 
 _T = TypeVar("_T")
+
+
+class GcovrDataAssertionError(AssertionError):
+    """Exception for data merge errors."""
+
+
+class GcovrMergeAssertionError(AssertionError):
+    """Exception for data merge errors."""
 
 
 def _presentable_filename(filename: str, root_filter: re.Pattern[str]) -> str:
@@ -146,9 +150,9 @@ class CoverageBase:
             )
         )
 
-    def raise_error(self, msg: str) -> NoReturn:
+    def raise_data_error(self, msg: str) -> NoReturn:
         """Get the exception with message extended with context."""
-        raise GcovrMergeAssertionError(
+        raise GcovrDataAssertionError(
             "\n".join(
                 [
                     msg,
@@ -228,7 +232,7 @@ class BranchCoverage(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if count < 0:
-            self.raise_error("count must not be a negative value.")
+            self.raise_data_error("count must not be a negative value.")
 
         self.branchno = branchno
         self.count = count
@@ -293,7 +297,7 @@ class BranchCoverage(CoverageBase):
         >>> left.merge(right, DEFAULT_MERGE_OPTIONS)
         Traceback (most recent call last):
           ...
-        gcovr.data_model.merging.GcovrMergeAssertionError: Source block ID must be equal, got 2 and 3.
+        gcovr.data_model.coverage.GcovrMergeAssertionError: Source block ID must be equal, got 2 and 3.
         GCOV data file of merge source is:
            right
         and of merge target is:
@@ -399,9 +403,9 @@ class ConditionCoverage(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if count < 0:
-            self.raise_error("count must not be a negative value.")
+            self.raise_data_error("count must not be a negative value.")
         if count < covered:
-            self.raise_error("count must not be less than covered.")
+            self.raise_data_error("count must not be less than covered.")
 
         self.conditionno = conditionno
         self.count = count
@@ -458,7 +462,7 @@ class ConditionCoverage(CoverageBase):
         >>> left.merge(right, DEFAULT_MERGE_OPTIONS)
         Traceback (most recent call last):
           ...
-        gcovr.data_model.merging.GcovrMergeAssertionError: The condition number must be equal, got 2 and expected 1.
+        gcovr.data_model.coverage.GcovrMergeAssertionError: The condition number must be equal, got 2 and expected 1.
         GCOV data file of merge source is:
            right
         and of merge target is:
@@ -590,10 +594,10 @@ class DecisionCoverageConditional(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if count_true < 0:
-            self.raise_error("count_true must not be a negative value.")
+            self.raise_data_error("count_true must not be a negative value.")
         self.count_true = count_true
         if count_false < 0:
-            self.raise_error("count_true must not be a negative value.")
+            self.raise_data_error("count_true must not be a negative value.")
         self.count_false = count_false
 
     def serialize(
@@ -666,7 +670,7 @@ class DecisionCoverageSwitch(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if count < 0:
-            self.raise_error("count must not be a negative value.")
+            self.raise_data_error("count must not be a negative value.")
         self.count = count
 
     def serialize(
@@ -788,7 +792,7 @@ class CallCoverage(CoverageBase):
         Do not use 'left' or 'right' objects afterwards!
         """
         if self.callno != other.callno:
-            self.raise_error(
+            self.raise_data_error(
                 f"Call number must be equal, got {self.callno} and {other.callno}."
             )
         self.covered |= other.covered
@@ -824,10 +828,14 @@ class FunctionCoverage(CoverageBase):
     to merge function coverage in different ways
 
     Args:
-        name (str):
-            The mangled name of the function, None if not available.
+        mangled_name (str):
+            The mangled name of the function. If demangled_name is None and
+            the name contains a brace it's used as demangled_name. This is needed
+            to support existing GCOV text output where we do not know if the
+            option --demanglednames was used for generation. If it contains a brace
+            the demangled name must be None.
         demangled_name (str):
-            The demangled name (signature) of the functions.
+            The demangled name of the functions.
         lineno (int):
             The line number.
         count (int):
@@ -843,7 +851,7 @@ class FunctionCoverage(CoverageBase):
     """
 
     __slots__ = (
-        "name",
+        "mangled_name",
         "demangled_name",
         "count",
         "blocks",
@@ -856,7 +864,7 @@ class FunctionCoverage(CoverageBase):
         self,
         data_source: Union[str, tuple[str, ...], set[tuple[str, ...]]],
         *,
-        name: Optional[str],
+        mangled_name: Optional[str],
         demangled_name: Optional[str],
         lineno: int,
         count: int,
@@ -867,8 +875,16 @@ class FunctionCoverage(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if count < 0:
-            self.raise_error("count must not be a negative value.")
-        self.name = name
+            self.raise_data_error("count must not be a negative value.")
+        if mangled_name is not None:
+            # We have a demangled name as name -> demangled_name must be None and we need to change the values
+            if "(" in mangled_name:
+                if demangled_name is not None:
+                    self.raise_data_error(
+                        f"If 'name' contains a demangled name (got '{mangled_name}') the 'demangled_name' must be None (got {demangled_name})."
+                    )
+                mangled_name, demangled_name = (None, mangled_name)
+        self.mangled_name = mangled_name
         self.demangled_name = demangled_name
         self.count = CoverageDict[int, int]({lineno: count})
         self.blocks = CoverageDict[int, float]({lineno: blocks})
@@ -890,8 +906,8 @@ class FunctionCoverage(CoverageBase):
         data_dict_list = list[dict[str, Any]]()
         for lineno, count in self.count.items():
             data_dict = dict[str, Any]()
-            if self.name is not None:
-                data_dict["name"] = self.name
+            if self.mangled_name is not None:
+                data_dict["name"] = self.mangled_name
             if self.demangled_name is not None:
                 data_dict["demangled_name"] = self.demangled_name
             data_dict.update(
@@ -928,8 +944,8 @@ class FunctionCoverage(CoverageBase):
 
         return FunctionCoverage(
             data_dict.get(GCOVR_DATA_SOURCES, data_source),
-            name=data_dict.get("name"),
-            demangled_name=data_dict["demangled_name"],
+            mangled_name=data_dict.get("name"),
+            demangled_name=data_dict.get("demangled_name"),
             lineno=data_dict["lineno"],
             count=data_dict["execution_count"],
             blocks=data_dict["blocks_percent"],
@@ -958,17 +974,38 @@ class FunctionCoverage(CoverageBase):
         - ``options.func_opts.merge_function_use_line_max``
         - ``options.func_opts.separate_function``
         """
-        self.name = self._merge_property(
-            other, "Function mangled name", lambda x: x.name
-        )
         self.demangled_name = self._merge_property(
             other, "Function demangled name", lambda x: x.demangled_name
         )
+        # If we have a demangled name use the first mangled name
+        # For virtual constructors/destructors several mangled functions map to the same demangled name,
+        # see https://itanium-cxx-abi.github.io/cxx-abi/abi.html#mangling-special-ctor-dtor:
+        # <ctor-dtor-name> ::= C1                     # complete object constructor
+        #                  ::= C2                     # base object constructor
+        #                  ::= C3                     # complete object allocating constructor
+        #                  ::= CI1 <base class type>  # complete object inheriting constructor
+        #                  ::= CI2 <base class type>  # base object inheriting constructor
+        #                  ::= D0                     # deleting destructor
+        #                  ::= D1                     # complete object destructor
+        #                  ::= D2                     # base object destructor
+        if self.demangled_name is not None:
+            if self.mangled_name is None:
+                self.mangled_name = other.mangled_name
+            elif (
+                other.mangled_name is not None
+                and other.mangled_name < self.mangled_name
+            ):
+                self.mangled_name = other.mangled_name
+        # If we do not have mangled names the mangled name must be the same.
+        else:
+            self.mangled_name = self._merge_property(
+                other, "Function mangled name", lambda x: x.mangled_name
+            )
         if not options.func_opts.ignore_function_lineno:
             if self.count.keys() != other.count.keys():
                 lines = sorted(set([*self.count.keys(), *other.count.keys()]))
                 self.raise_merge_error(
-                    f"Got function {self.demangled_name} on multiple lines: {', '.join([str(line) for line in lines])}.\n"
+                    f"Got function {self.name} on multiple lines: {', '.join([str(line) for line in lines])}.\n"
                     "\tYou can run gcovr with --merge-mode-functions=MERGE_MODE.\n"
                     "\tThe available values for MERGE_MODE are described in the documentation.",
                     other,
@@ -1047,7 +1084,40 @@ class FunctionCoverage(CoverageBase):
     @property
     def key(self) -> str:
         """Get the key for the dict."""
-        return str(self.name or self.demangled_name)
+        return self.name
+
+    @property
+    def name(self) -> str:
+        """Get the function name. This is the demangled name if present, else the mangled name."""
+        return str(self.demangled_name or self.mangled_name)
+
+    def is_function(self, name: Optional[str]) -> bool:
+        """Is the name one of the function."""
+        return name is not None and (name in (self.mangled_name, self.demangled_name))
+
+    @property
+    def name_and_signature(self) -> tuple[str, str]:
+        """Get a tuple with function name and signature, if signature is un."""
+        if self.demangled_name is None:
+            return (str(self.name), "")
+
+        if "(" not in self.demangled_name:
+            return (str(self.demangled_name), "")
+
+        open_brackets, close_brackets = (0, 0)
+        signature = ""
+        for part in reversed(self.demangled_name.split("(")):
+            signature = f"({part}{signature}"
+            open_brackets += 1
+            close_brackets += len(re.findall(r"(\))", part))
+            if open_brackets == close_brackets:
+                break
+        else:
+            self.raise_data_error(
+                f"Can't split function {self.demangled_name!r} into name and signature."
+            )
+
+        return (self.demangled_name[: -len(signature)], signature)
 
 
 class LineCoverage(CoverageBase):
@@ -1100,9 +1170,9 @@ class LineCoverage(CoverageBase):
     ) -> None:
         super().__init__(data_source)
         if lineno <= 0:
-            self.raise_error("Line number must be a positive value.")
+            self.raise_data_error("Line number must be a positive value.")
         if count < 0:
-            self.raise_error("count must not be a negative value.")
+            self.raise_data_error("count must not be a negative value.")
 
         self.lineno = lineno
         self.count = count
@@ -1427,7 +1497,7 @@ class FileCoverage(CoverageBase):
         """
 
         if self.filename != other.filename:
-            self.raise_error("Filename must be equal")
+            self.raise_data_error("Filename must be equal")
 
         self.lines.merge(other.lines, options)
         self.functions.merge(other.functions, options)
@@ -1559,7 +1629,7 @@ class FileCoverage(CoverageBase):
     def filter_for_function(self, functioncov: FunctionCoverage) -> FileCoverage:
         """Get a file coverage object reduced to a single function"""
         if functioncov.key not in self.functions:
-            self.raise_error(
+            self.raise_data_error(
                 f"Function {functioncov.key} must be in filtered file coverage object."
             )
         filecov = FileCoverage(self.data_sources, filename=self.filename)
@@ -1569,8 +1639,7 @@ class FileCoverage(CoverageBase):
             {
                 key: linecov
                 for key, linecov in self.lines.items()
-                if linecov.function_name
-                == (functioncov.name or functioncov.demangled_name)
+                if functioncov.is_function(linecov.function_name)
             }
         )