+++ src/gcovr/exclusions/__init__.py
@@ -146,10 +146,12 @@ def remove_internal_functions(filecov: FileCoverage) -> None:
     # Get all the keys first because we want to remove some of them which will else result in an error.
     for key in list(filecov.functions.keys()):
         functioncov = filecov.functions[key]
-        if _function_can_be_excluded(str(functioncov.demangled_name)):
+        if _function_can_be_excluded(
+            functioncov.mangled_name, functioncov.demangled_name
+        ):
             LOGGER.debug(
                 "Ignoring symbol %s in line %s in file %s",
-                functioncov.demangled_name,
+                functioncov.name,
                 ", ".join([str(line) for line in sorted(functioncov.count.keys())]),
                 filecov.filename,
             )
@@ -157,16 +159,17 @@ def remove_internal_functions(filecov: FileCoverage) -> None:
             # Remove function and exclude the related lines
             filecov.functions.pop(key)
             for linecov in filecov.lines.values():
-                if (
-                    linecov.function_name is not None
-                    and linecov.function_name == functioncov.name
-                ):
+                if functioncov.is_function(linecov.function_name):
                     linecov.exclude()
 
 
-def _function_can_be_excluded(name: str) -> bool:
+def _function_can_be_excluded(*names: Optional[str]) -> bool:
     """special names for construction/destruction of static objects will be ignored"""
-    return name.startswith("__") or name.startswith("_GLOBAL__sub_I_")
+    return any(
+        name is not None
+        and (name.startswith("__") or name.startswith("_GLOBAL__sub_I_"))
+        for name in names
+    )
 
 
 def remove_function_lines(filecov: FileCoverage) -> None:
@@ -203,16 +206,26 @@ def remove_functions(filecov: FileCoverage, patterns: list[re.Pattern[str]]) -> 
 
         exclude_ranges = []
         for lineno, functions in functions_by_line.items():
-            for function in functions:
+            for functioncov in functions:
                 for pattern in patterns:
-                    if pattern.fullmatch(str(function.demangled_name)):
-                        if function.start is None or function.start[lineno] is None:
+                    if any(
+                        pattern.fullmatch(name)
+                        for name in (
+                            functioncov.name,
+                            functioncov.demangled_name,
+                        )
+                        if name is not None
+                    ):
+                        if (
+                            functioncov.start is None
+                            or functioncov.start[lineno] is None
+                        ):
                             function_exclude_not_supported()
                         else:
                             exclude_ranges += get_function_exclude_ranges(
                                 filecov.filename,
                                 lineno,
-                                function.start[lineno][1]
+                                functioncov.start[lineno][1]
                                 + 1,  # Cheat that the comment is after the definition
                                 functions_by_line=functions_by_line,
                             )