+++ src/gcovr/formats/html/write.py
@@ -825,17 +825,19 @@ def get_file_data(
     )
     functions = dict[tuple[str, str, int], dict[str, Any]]()
     # Only use demangled names (containing a brace)
-    for f_cdata in sorted(cdata.functions.values(), key=lambda f_cdata: f_cdata.key):
-        for lineno in sorted(f_cdata.count.keys()):
+    for functioncov in sorted(
+        cdata.functions.values(), key=lambda functioncov: functioncov.key
+    ):
+        for lineno in sorted(functioncov.count.keys()):
             f_data = dict[str, Any]()
-            f_data["name"] = f_cdata.demangled_name
+            f_data["name"] = functioncov.name
             f_data["filename"] = cdata_fname[filename]
             f_data["html_filename"] = os.path.basename(cdata_sourcefile[filename])
             f_data["line"] = lineno
-            f_data["count"] = f_cdata.count[lineno]
-            f_data["blocks"] = f_cdata.blocks[lineno]
-            f_data["excluded"] = f_cdata.excluded[lineno]
-            function_stats = cdata.filter_for_function(f_cdata).stats
+            f_data["count"] = functioncov.count[lineno]
+            f_data["blocks"] = functioncov.blocks[lineno]
+            f_data["excluded"] = functioncov.excluded[lineno]
+            function_stats = cdata.filter_for_function(functioncov).stats
             f_data["line_coverage"] = function_stats.line.percent_or(100.0)
             f_data["branch_coverage"] = function_stats.branch.percent_or("-")
             f_data["condition_coverage"] = function_stats.condition.percent_or("-")
@@ -843,7 +845,7 @@ def get_file_data(
             file_data["function_list"].append(f_data)
             functions[
                 (
-                    f_cdata.key,
+                    functioncov.key,
                     str(f_data["filename"]),
                     int(f_data["line"]),
                 )